pyproject.toml
[tool.poetry]
name = "dianbot"
version = "0.1.0"
description = ""
authors = ["dianhsu <xudian.cn@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
khl-py = "^0.3.16"
loguru = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
